requirements.txt
# RAG系统核心依赖
requests>=2.31.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# 文档处理
python-docx>=0.8.11
PyPDF2>=3.0.1
beautifulsoup4>=4.12.0
unstructured>=0.10.0

# 向量化和模型
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Web框架（可选，用于API服务）
fastapi>=0.100.0
uvicorn>=0.23.0

# 数据库
chromadb>=0.4.0
sqlite3

# 工具库
textwrap
json
